requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-ipware==7.0.1
flake8==7.1.2
idna==3.10
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
python-ipware==3.0.0
requests==2.32.3
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
urllib3==2.3.0
